+++ src/module.js
@@ -21,6 +21,7 @@ import registerLibwrappers from "./libwrapper.js";
 import { DatabaseViewerApp } from "./formapplications/database/database-viewer-app.js";
 import { EffectsUIApp } from "./formapplications/effects-ui/effects-ui-app.js";
 import CONSTANTS from "./constants.js";
+import { PlayerSettings } from "./formapplications/effects-ui/effect-player-store.js";
 
 Hooks.once('init', async function () {
   if (!game.modules.get("socketlib")?.active) return;
@@ -38,6 +39,8 @@ Hooks.once('ready', async function () {
     ui.notifications.error("Sequencer requires the SocketLib module to be active and will not work without it!", { console: false });
     throw new Error("Sequencer requires the SocketLib module to be active and will not work without it!");
   }
+
+  PlayerSettings.migrateOldPresets();
 
   InteractionManager.initialize();