+++ src/modules/sequencer-pixi-animation-helper.js
@@ -1,0 +1,12 @@
+import * as lib from "../lib/lib.js";
+import * as canvaslib from "../lib/canvas-lib.js";
+import SequencerAnimationEngine from "./sequencer-animation-engine.js";
+
+export default class SequencerPIXIAnimationHelper {
+
+  static play(data){
+
+
+  }
+
+}